Added assert to ensure that the maximum rank is equal to the number of merged samples

+++ include/idr.h
@@ -27,7 +27,7 @@ struct OptimizationRV
 
 struct OptimizationRV
 em_gaussian(
-    size_t n_samples,
+    int n_samples,
     double* x, 
     double* y,
     double* localIDR);
